+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+- **Arrays**
+  - Add the `merge()` function with the `MergeOption` and `NullsOption` helpers.
 
 ## [1.0.7] - 2025-12-12